+++ test/test_helper.rb
@@ -2,6 +2,7 @@ ENV["RAILS_ENV"] ||= "test"
 require_relative "../config/environment"
 require_relative "support/unifi_api_stubs"
 require_relative "support/mailersend_api_stubs"
+require_relative "support/smsapi_stubs"
 require "rails/test_help"
 require "webmock/minitest"
 require "minitest/mock"
@@ -18,5 +19,6 @@ module ActiveSupport
     # Add more helper methods to be used by all tests here...
     include UnifiApiStubs
     include MailersendApiStubs
+    include SmsapiStubs
   end
 end